+++ server/libs/modules/task-dispatchers/forkjoin/src/main/java/com/bytechef/task/dispatcher/forkjoin/ForkJoinTaskDispatcherDefinitionFactory.java
@@ -15,14 +15,14 @@
  * limitations under the License.
  */
 
-package com.bytechef.task.dispatcher.fork;
+package com.bytechef.task.dispatcher.forkjoin;
 
 import static com.bytechef.hermes.task.dispatcher.definition.TaskDispatcherDSL.array;
 import static com.bytechef.hermes.task.dispatcher.definition.TaskDispatcherDSL.display;
 import static com.bytechef.hermes.task.dispatcher.definition.TaskDispatcherDSL.task;
 import static com.bytechef.hermes.task.dispatcher.definition.TaskDispatcherDSL.taskDispatcher;
-import static com.bytechef.task.dispatcher.fork.constants.ForkJoinTaskDispatcherConstants.BRANCHES;
-import static com.bytechef.task.dispatcher.fork.constants.ForkJoinTaskDispatcherConstants.FORK_JOIN;
+import static com.bytechef.task.dispatcher.forkjoin.constants.ForkJoinTaskDispatcherConstants.BRANCHES;
+import static com.bytechef.task.dispatcher.forkjoin.constants.ForkJoinTaskDispatcherConstants.FORK_JOIN;
 
 import com.bytechef.hermes.task.dispatcher.TaskDispatcherDefinitionFactory;
 import com.bytechef.hermes.task.dispatcher.definition.TaskDispatcherDefinition;
