+++ server/libs/modules/task-dispatchers/forkjoin/src/main/java/com/bytechef/task/dispatcher/forkjoin/constants/ForkJoinTaskDispatcherConstants.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.bytechef.task.dispatcher.fork.constants;
+package com.bytechef.task.dispatcher.forkjoin.constants;
 
 /**
  * @author Ivica Cardic
