+++ server/libs/modules/task-dispatchers/forkjoin/src/test/java/com/bytechef/task/dispatcher/forkjoin/ForkJoinTaskDispatcherDefinitionFactoryTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.bytechef.task.dispatcher.fork;
+package com.bytechef.task.dispatcher.forkjoin;
 
 import com.bytechef.test.jsonasssert.JsonFileAssert;
 import org.junit.jupiter.api.Test;
