+++ server/libs/modules/task-dispatchers/forkjoin/src/test/resources/definition/fork-join_v1.json
