+++ server/libs/modules/task-dispatchers/loop/src/test/java/com/bytechef/task/dispatcher/loop/LoopTaskDispatcherDefinitionFactoryTest.java
@@ -27,6 +27,7 @@ public class LoopTaskDispatcherDefinitionFactoryTest {
 
     @Test
     public void testGetTaskDispatcherDefinition() {
-        JsonFileAssert.assertEquals("definition/loop_v1.json", new LoopTaskDispatcherDefinitionFactory().getDefinition());
+        JsonFileAssert.assertEquals("definition/loop_v1.json",
+            new LoopTaskDispatcherDefinitionFactory().getDefinition());
     }
 }
