+++ server/libs/modules/task-dispatchers/parallel/src/test/java/com/bytechef/task/dispatcher/parallel/ParallelTaskDispatcherDefinitionFactoryTest.java
@@ -27,6 +27,7 @@ public class ParallelTaskDispatcherDefinitionFactoryTest {
 
     @Test
     public void testGetTaskDispatcherDefinition() {
-        JsonFileAssert.assertEquals("definition/parallel_v1.json", new ParallelTaskDispatcherDefinitionFactory().getDefinition());
+        JsonFileAssert.assertEquals("definition/parallel_v1.json",
+            new ParallelTaskDispatcherDefinitionFactory().getDefinition());
     }
 }
