+++ server/libs/modules/task-dispatchers/sequence/src/test/java/com/bytechef/task/dispatcher/sequence/SequenceTaskDispatcherDefinitionFactoryTest.java
@@ -27,6 +27,7 @@ public class SequenceTaskDispatcherDefinitionFactoryTest {
 
     @Test
     public void testGetTaskDispatcherDefinition() {
-        JsonFileAssert.assertEquals("definition/sequence_v1.json", new SequenceTaskDispatcherDefinitionFactory().getDefinition());
+        JsonFileAssert.assertEquals("definition/sequence_v1.json",
+            new SequenceTaskDispatcherDefinitionFactory().getDefinition());
     }
 }
