+++ settings.gradle
@@ -160,7 +160,7 @@ include("server:libs:modules:components:xlsxfile")
 include("server:libs:modules:components:xmlfile")
 include("server:libs:modules:components:xmlhelper")
 include("server:libs:modules:task-dispatchers:each")
-include("server:libs:modules:task-dispatchers:fork")
+include("server:libs:modules:task-dispatchers:forkjoin")
 include("server:libs:modules:task-dispatchers:if")
 include("server:libs:modules:task-dispatchers:loop")
 include("server:libs:modules:task-dispatchers:map")
